requirements.txt
streamlit==1.32.0
groq==0.7.0
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.2.5
pydantic==1.10.12
reportlab==4.1.5
requests==2.32.0
pytest==7.4.2
